# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API Backend NeoPay - Sistem Antrian Digital Bank
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  pesan: string;
  error?: string;
}

export interface PesanResponse {
  pesan: string;
}

export interface DaftarBody {
  nama: string;
  email: string;
  /** @minLength 6 */
  password: string;
  noHp: string;
  nik?: string | null;
  /** Tanggal lahir format YYYY-MM-DD */
  tanggalLahir: string;
}

export interface MasukBody {
  email: string;
  password: string;
}

export type NasabahResponseRole =
  (typeof NasabahResponseRole)[keyof typeof NasabahResponseRole];

export const NasabahResponseRole = {
  nasabah: "nasabah",
  teller: "teller",
  admin: "admin",
} as const;

export interface NasabahResponse {
  id: string;
  supabaseUserId: string;
  nama: string;
  noHp: string;
  email: string;
  nik?: string | null;
  noRekening?: string | null;
  tanggalLahir?: string | null;
  /** Saldo nasabah dalam rupiah */
  saldo: number;
  role: NasabahResponseRole;
  oneSignalPlayerId?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  nasabah: NasabahResponse;
}

export type AntrianPublikItemJenisLayanan =
  (typeof AntrianPublikItemJenisLayanan)[keyof typeof AntrianPublikItemJenisLayanan];

export const AntrianPublikItemJenisLayanan = {
  teller: "teller",
  cs: "cs",
  kredit: "kredit",
} as const;

export type AntrianPublikItemStatus =
  (typeof AntrianPublikItemStatus)[keyof typeof AntrianPublikItemStatus];

export const AntrianPublikItemStatus = {
  menunggu: "menunggu",
  dipanggil: "dipanggil",
  sedang_dilayani: "sedang_dilayani",
  selesai: "selesai",
  dibatalkan: "dibatalkan",
} as const;

/**
 * Data antrian publik tanpa PII (untuk tampilan layar tunggu)
 */
export interface AntrianPublikItem {
  id: string;
  tanggal: string;
  nomorAntrian: number;
  jenisLayanan: AntrianPublikItemJenisLayanan;
  status: AntrianPublikItemStatus;
  createdAt: string;
  updatedAt: string;
}

export type AntrianItemJenisLayanan =
  (typeof AntrianItemJenisLayanan)[keyof typeof AntrianItemJenisLayanan];

export const AntrianItemJenisLayanan = {
  teller: "teller",
  cs: "cs",
  kredit: "kredit",
} as const;

export type AntrianItemStatus =
  (typeof AntrianItemStatus)[keyof typeof AntrianItemStatus];

export const AntrianItemStatus = {
  menunggu: "menunggu",
  dipanggil: "dipanggil",
  sedang_dilayani: "sedang_dilayani",
  selesai: "selesai",
  dibatalkan: "dibatalkan",
} as const;

/**
 * Data antrian lengkap dengan nasabahId (untuk nasabah atau admin yang terautentikasi)
 */
export interface AntrianItem {
  id: string;
  nasabahId: string;
  cabangId?: string | null;
  tanggal: string;
  nomorAntrian: number;
  jenisLayanan: AntrianItemJenisLayanan;
  status: AntrianItemStatus;
  catatan?: string | null;
  dilayaniOlehId?: string | null;
  createdAt: string;
  updatedAt: string;
  namaNasabah?: string | null;
  noHpNasabah?: string | null;
  namaCabang?: string | null;
  alamatCabang?: string | null;
  kodeCabang?: string | null;
  namaPetugas?: string | null;
}

export interface AntrianDetailResponse {
  antrian: AntrianItem;
  posisi?: number | null;
}

export interface StatusAntrianResponse {
  antrian: AntrianItem;
  posisi: number;
  totalMenunggu: number;
  dipanggilSekarang?: AntrianPublikItem | null;
}

/**
 * Daftar antrian publik tanpa PII (untuk /antrian/hari-ini)
 */
export interface DaftarAntrianPublikResponse {
  data: AntrianPublikItem[];
  total: number;
}

/**
 * Daftar antrian lengkap dengan nasabahId (untuk endpoint terautentikasi)
 */
export interface DaftarAntrianResponse {
  data: AntrianItem[];
  total: number;
}

export type AmbilAntrianBodyJenisLayanan =
  (typeof AmbilAntrianBodyJenisLayanan)[keyof typeof AmbilAntrianBodyJenisLayanan];

export const AmbilAntrianBodyJenisLayanan = {
  teller: "teller",
  cs: "cs",
  kredit: "kredit",
} as const;

export interface AmbilAntrianBody {
  jenisLayanan?: AmbilAntrianBodyJenisLayanan;
  catatan?: string;
}

export interface JadwalResponse {
  id: string;
  hariBuka: string[];
  jamMulai: string;
  jamSelesai: string;
  kuotaPerHari: number;
  aktif: boolean;
  keterangan?: string | null;
}

export interface UpdateJadwalBody {
  hariBuka?: string[];
  jamMulai?: string;
  jamSelesai?: string;
  kuotaPerHari?: number;
  aktif?: boolean;
  keterangan?: string;
}

export interface UpdateProfilBody {
  nama?: string;
  noHp?: string;
}

export interface DaftarNotifBody {
  playerId: string;
}

export interface DasborResponse {
  tanggal: string;
  totalAntrian: number;
  menunggu: number;
  dipanggil: number;
  sedangDilayani: number;
  selesai: number;
  dibatalkan: number;
  tellerCount?: number;
  csCount?: number;
  kreditCount?: number;
  antrianAktif?: AntrianItem | null;
}

export interface StatusWAResponse {
  terhubung: boolean;
  nomorWA?: string | null;
  status: string;
}

export interface QrWAResponse {
  qr: string;
  pesan: string;
}