# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API Backend NeoPay - Sistem Antrian Digital Bank
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/antrian/{id}/panggil:
    put:
      operationId: panggilAntrian
      tags: [admin]
      summary: Panggil nasabah
      description: Admin/Teller memanggil nomor antrian berikutnya
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Nasabah berhasil dipanggil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AntrianDetailResponse"
        "404":
          description: Antrian tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/antrian/{id}/selesai:
    put:
      operationId: selesaiAntrian
      tags: [admin]
      summary: Tandai selesai
      description: Admin menandai nasabah sudah selesai dilayani
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Antrian ditandai selesai
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AntrianDetailResponse"
        "404":
          description: Antrian tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/antrian/{id}/skip:
    put:
      operationId: skipAntrian
      tags: [admin]
      summary: Skip nasabah
      description: Admin melewati nasabah (menandai dibatalkan)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Antrian dilewati
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AntrianDetailResponse"
        "404":
          description: Antrian tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/dasbor:
    get:
      operationId: getDasbor
      tags: [admin]
      summary: Statistik dasbor
      description: Statistik antrian hari ini untuk admin
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Statistik dasbor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DasborResponse"
...
  /admin/jadwal:
    put:
      operationId: updateJadwal
      tags: [admin]
      summary: Update jadwal
      description: Admin memperbarui jadwal operasional bank
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJadwalBody"
      responses:
        "200":
          description: Jadwal berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JadwalResponse"
...
  /admin/wa/status:
    get:
      operationId: getStatusWA
      tags: [whatsapp]
      summary: Status WhatsApp
      description: Mengecek status koneksi WhatsApp
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Status WhatsApp
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusWAResponse"
  /admin/wa/qr:
    get:
      operationId: getQrWA
      tags: [whatsapp]
      summary: QR Code WhatsApp
      description: Mendapatkan QR code untuk scan WhatsApp
      security:
        - bearerAuth: []
      responses:
        "200":
          description: QR code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrWAResponse"
        "409":
          description: WhatsApp sudah terhubung
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        pesan:
          type: string
        error:
          type: string
      required:
        - pesan
...
    AntrianItem:
      type: object
      description: Data antrian lengkap dengan nasabahId (untuk nasabah atau admin yang terautentikasi)
      properties:
        id:
          type: string
        nasabahId:
          type: string
        cabangId:
          type: string
          nullable: true
        tanggal:
          type: string
        nomorAntrian:
          type: integer
        jenisLayanan:
          type: string
          enum: [teller, cs, kredit]
        status:
          type: string
          enum: [menunggu, dipanggil, sedang_dilayani, selesai, dibatalkan]
        catatan:
          type: string
          nullable: true
        dilayaniOlehId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        namaNasabah:
          type: string
          nullable: true
        noHpNasabah:
          type: string
          nullable: true
        namaCabang:
          type: string
          nullable: true
        alamatCabang:
          type: string
          nullable: true
        kodeCabang:
          type: string
          nullable: true
        namaPetugas:
          type: string
          nullable: true
      required:
        - id
        - nasabahId
        - tanggal
        - nomorAntrian
        - jenisLayanan
        - status
        - createdAt
        - updatedAt
    AntrianDetailResponse:
      type: object
      properties:
        antrian:
          $ref: "#/components/schemas/AntrianItem"
        posisi:
          type: integer
          nullable: true
      required:
        - antrian
...
    JadwalResponse:
      type: object
      properties:
        id:
          type: string
        hariBuka:
          type: array
          items:
            type: string
        jamMulai:
          type: string
        jamSelesai:
          type: string
        kuotaPerHari:
          type: integer
        aktif:
          type: boolean
        keterangan:
          type: string
          nullable: true
      required:
        - id
        - hariBuka
        - jamMulai
        - jamSelesai
        - kuotaPerHari
        - aktif
    UpdateJadwalBody:
      type: object
      properties:
        hariBuka:
          type: array
          items:
            type: string
        jamMulai:
          type: string
        jamSelesai:
          type: string
        kuotaPerHari:
          type: integer
        aktif:
          type: boolean
        keterangan:
          type: string
...
    DasborResponse:
      type: object
      properties:
        tanggal:
          type: string
        totalAntrian:
          type: integer
        menunggu:
          type: integer
        dipanggil:
          type: integer
        sedangDilayani:
          type: integer
        selesai:
          type: integer
        dibatalkan:
          type: integer
        tellerCount:
          type: integer
        csCount:
          type: integer
        kreditCount:
          type: integer
        antrianAktif:
          $ref: "#/components/schemas/AntrianItem"
          nullable: true
      required:
        - tanggal
        - totalAntrian
        - menunggu
        - dipanggil
        - sedangDilayani
        - selesai
        - dibatalkan
    StatusWAResponse:
      type: object
      properties:
        terhubung:
          type: boolean
        nomorWA:
          type: string
          nullable: true
        status:
          type: string
      required:
        - terhubung
        - status
    QrWAResponse:
      type: object
      properties:
        qr:
          type: string
        pesan:
          type: string
      required:
        - qr
        - pesan