# lib/api-zod/src/generated/types/antrianDetailResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API Backend NeoPay - Sistem Antrian Digital Bank
 * OpenAPI spec version: 0.1.0
 */
import type { AntrianItem } from "./antrianItem";

export interface AntrianDetailResponse {
  antrian: AntrianItem;
  posisi?: number | null;
}