# lib/api-zod/src/generated/types/antrianItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API Backend NeoPay - Sistem Antrian Digital Bank
 * OpenAPI spec version: 0.1.0
 */
import type { AntrianItemJenisLayanan } from "./antrianItemJenisLayanan";
import type { AntrianItemStatus } from "./antrianItemStatus";

/**
 * Data antrian lengkap dengan nasabahId (untuk nasabah atau admin yang terautentikasi)
 */
export interface AntrianItem {
  id: string;
  nasabahId: string;
  cabangId?: string | null;
  tanggal: string;
  nomorAntrian: number;
  jenisLayanan: AntrianItemJenisLayanan;
  status: AntrianItemStatus;
  catatan?: string | null;
  dilayaniOlehId?: string | null;
  createdAt: Date;
  updatedAt: Date;
  namaNasabah?: string | null;
  noHpNasabah?: string | null;
  namaCabang?: string | null;
  alamatCabang?: string | null;
  kodeCabang?: string | null;
  namaPetugas?: string | null;
}