# lib/api-zod/src/generated/types/qrWAResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API Backend NeoPay - Sistem Antrian Digital Bank
 * OpenAPI spec version: 0.1.0
 */

export interface QrWAResponse {
  qr: string;
  pesan: string;
}