# lib/api-zod/src/generated/types/statusWAResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API Backend NeoPay - Sistem Antrian Digital Bank
 * OpenAPI spec version: 0.1.0
 */

export interface StatusWAResponse {
  terhubung: boolean;
  nomorWA?: string | null;
  status: string;
}